[CHORE] Extension & Struct 정리
